+++ packages/houdini/src/codegen/utils/flattenSelections.ts
@@ -10,6 +10,7 @@ export function flattenSelections({
 	fragmentDefinitions,
 	ignoreMaskDisable,
 	keepFragmentSpreadNodes,
+	hoistFragments,
 }: {
 	config: Config
 	filepath: string
@@ -17,6 +18,7 @@ export function flattenSelections({
 	fragmentDefinitions: { [name: string]: graphql.FragmentDefinitionNode }
 	ignoreMaskDisable?: boolean
 	keepFragmentSpreadNodes?: boolean
+	hoistFragments?: boolean
 }): readonly graphql.SelectionNode[] {
 	// collect all of the fields together
 	const fields = new FieldCollection({
@@ -26,6 +28,7 @@ export function flattenSelections({
 		fragmentDefinitions,
 		ignoreMaskDisable: !!ignoreMaskDisable,
 		keepFragmentSpreadNodes: !!keepFragmentSpreadNodes,
+		hoistFragments,
 	})
 
 	// convert the flat fields into a selection set
@@ -42,6 +45,7 @@ class FieldCollection {
 	fragmentSpreads: { [fragmentName: string]: graphql.FragmentSpreadNode }
 	ignoreMaskDisable: boolean
 	keepFragmentSpreadNodes: boolean
+	hoistFragments: boolean
 
 	constructor(args: {
 		config: Config
@@ -50,6 +54,7 @@ class FieldCollection {
 		fragmentDefinitions: { [name: string]: graphql.FragmentDefinitionNode }
 		ignoreMaskDisable: boolean
 		keepFragmentSpreadNodes: boolean
+		hoistFragments?: boolean
 	}) {
 		this.config = args.config
 		this.fragmentDefinitions = args.fragmentDefinitions
@@ -59,11 +64,12 @@ class FieldCollection {
 		this.fields = {}
 		this.inlineFragments = {}
 		this.fragmentSpreads = {}
+		this.hoistFragments = !!args.hoistFragments
 
 		this.filepath = args.filepath
 
 		for (const selection of args.selections) {
-			this.add(selection)
+			this.add({ selection })
 		}
 	}
 
@@ -75,7 +81,36 @@ class FieldCollection {
 		)
 	}
 
-	add(selection: graphql.SelectionNode) {
+	add({ selection, external }: { selection: graphql.SelectionNode; external?: boolean }) {
+		// we need to figure out if we want to include this fragment's selection in
+		// the final result.
+
+		// the default behavior depends on wether masking is enabled or disabled
+		let includeFragments = this.config.defaultFragmentMasking === 'disable'
+
+		// Check if locally enable
+		const maskEnableDirective = selection.directives?.find(
+			({ name }) => name.value === this.config.maskEnableDirective
+		)
+		if (maskEnableDirective) {
+			includeFragments = false
+		}
+
+		// Check if locally disable
+		const maskDisableDirective = selection.directives?.find(
+			({ name }) => name.value === this.config.maskDisableDirective
+		)
+		if (maskDisableDirective) {
+			includeFragments = true
+		}
+
+		// we might need to ignore any disables
+		// for example, queries need to _always_ have their full selection
+		// so the result can be written to the cache
+		if (this.ignoreMaskDisable) {
+			includeFragments = true
+		}
+
 		// how we handle the field depends on what kind of field it is
 		if (selection.kind === 'Field') {
 			// figure out the key of the field
@@ -92,14 +127,19 @@ class FieldCollection {
 
 			// its safe to all this fields selections if they exist
 			for (const subselect of selection.selectionSet?.selections || []) {
-				this.fields[key].selection.add(subselect)
+				this.fields[key].selection.add({
+					selection: subselect,
+					external,
+				})
 			}
 
 			// the application of the fragment has been validated already so track it
 			// so we can recreate
-			this.fields[key].selection.fragmentSpreads = {
-				...this.collectFragmentSpreads(selection.selectionSet?.selections ?? []),
-				...this.fields[key].selection.fragmentSpreads,
+			if (!external && includeFragments) {
+				this.fields[key].selection.fragmentSpreads = {
+					...this.collectFragmentSpreads(selection.selectionSet?.selections ?? []),
+					...this.fields[key].selection.fragmentSpreads,
+				}
 			}
 
 			// we're done
@@ -109,7 +149,7 @@ class FieldCollection {
 		// we could run into an inline fragment that doesn't assert a type (treat it as a field selection)
 		if (selection.kind === 'InlineFragment' && !selection.typeCondition) {
 			for (const subselect of selection.selectionSet.selections) {
-				this.add(subselect)
+				this.add({ selection: subselect, external })
 			}
 		}
 
@@ -119,7 +159,7 @@ class FieldCollection {
 			// we need to look at the selection set flatten all of the nested
 			// inline fragments that could appear. every time we run into a new
 			// inline fragment we need to create a new key in the object to add fields to
-			this.walkInlineFragment(selection)
+			this.walkInlineFragment({ selection, external, hoistFragments: this.hoistFragments })
 
 			// we're done
 			return
@@ -127,37 +167,8 @@ class FieldCollection {
 
 		// the only thing that's left is external fragment spreads
 		if (selection.kind === 'FragmentSpread') {
-			// we need to figure out if we want to include this fragment's selection in
-			// the final result.
-
-			// the default behavior depends on wether masking is enabled or disabled
-			let includeFragments = this.config.defaultFragmentMasking === 'disable'
-
-			// Check if locally enable
-			const maskEnableDirective = selection.directives?.find(
-				({ name }) => name.value === this.config.maskEnableDirective
-			)
-			if (maskEnableDirective) {
-				includeFragments = false
-			}
-
-			// Check if locally disable
-			const maskDisableDirective = selection.directives?.find(
-				({ name }) => name.value === this.config.maskDisableDirective
-			)
-			if (maskDisableDirective) {
-				includeFragments = true
-			}
-
-			// we might need to ignore any disables
-			// for example, queries need to _always_ have their full selection
-			// so the result can be written to the cache
-			if (this.ignoreMaskDisable) {
-				includeFragments = true
-			}
-
 			// make sure we leave this fragment in the selection behind
-			if (this.keepFragmentSpreadNodes) {
+			if (this.keepFragmentSpreadNodes && !external) {
 				this.fragmentSpreads[selection.name.value] = selection
 			}
 
@@ -181,19 +192,23 @@ class FieldCollection {
 			}
 
 			// instead of adding the field on directly, let's turn the external fragment into an inline fragment
+			// and process it like normal
 			this.add({
-				kind: 'InlineFragment',
-				typeCondition: {
-					kind: 'NamedType',
-					name: {
-						kind: 'Name',
-						value: definition.typeCondition.name.value,
+				selection: {
+					kind: 'InlineFragment',
+					typeCondition: {
+						kind: 'NamedType',
+						name: {
+							kind: 'Name',
+							value: definition.typeCondition.name.value,
+						},
+					},
+					selectionSet: {
+						kind: 'SelectionSet',
+						selections: [...definition.selectionSet.selections],
 					},
 				},
-				selectionSet: {
-					kind: 'SelectionSet',
-					selections: [...definition.selectionSet.selections],
-				},
+				external,
 			})
 		}
 	}
@@ -278,7 +293,15 @@ class FieldCollection {
 			)
 	}
 
-	walkInlineFragment(selection: graphql.InlineFragmentNode) {
+	walkInlineFragment({
+		selection,
+		external,
+		hoistFragments,
+	}: {
+		selection: graphql.InlineFragmentNode
+		external?: boolean
+		hoistFragments: boolean
+	}) {
 		// figure out the key of the field
 		const key = selection.typeCondition!.name.value
 
@@ -295,14 +318,32 @@ class FieldCollection {
 		for (const subselect of selection.selectionSet.selections || []) {
 			// the only thing we need to treat specially is inline fragments with type conditions,
 			// otherwise just add it like normal to the inline fragment
-			if (subselect.kind !== 'InlineFragment' || !subselect.typeCondition) {
-				this.inlineFragments[key].selection.add(subselect)
+			if (
+				(subselect.kind === 'FragmentSpread' && !hoistFragments) ||
+				subselect.kind === 'Field' ||
+				(subselect.kind === 'InlineFragment' && !subselect.typeCondition)
+			) {
+				this.inlineFragments[key].selection.add({
+					selection: subselect,
+					external,
+				})
+				continue
+			}
+
+			//  its a fragment spread that we should mix into the parent
+			else if (subselect.kind === 'FragmentSpread') {
+				this.add({
+					selection: subselect,
+					external: true,
+				})
 				continue
 			}
 
 			// we know the selection is an inline fragment with a type condition so
 			// flatten the selection into this one
-			this.walkInlineFragment(subselect)
+			else {
+				this.walkInlineFragment({ selection: subselect, external, hoistFragments })
+			}
 		}
 	}
 
@@ -314,6 +355,7 @@ class FieldCollection {
 			filepath: this.filepath,
 			ignoreMaskDisable: this.ignoreMaskDisable,
 			keepFragmentSpreadNodes: this.keepFragmentSpreadNodes,
+			hoistFragments: this.hoistFragments,
 		})
 	}
 }